+++ cpp/dolfinx/io/CMakeLists.txt
@@ -5,6 +5,7 @@ set(HEADERS_io
   ${CMAKE_CURRENT_SOURCE_DIR}/HDF5Interface.h
   ${CMAKE_CURRENT_SOURCE_DIR}/pugiconfig.hpp
   ${CMAKE_CURRENT_SOURCE_DIR}/pugixml.hpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/vtk_utils.h
   ${CMAKE_CURRENT_SOURCE_DIR}/VTKFile.h
   ${CMAKE_CURRENT_SOURCE_DIR}/XDMFFile.h
   ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_function.h
@@ -20,6 +21,7 @@ target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/HDF5Interface.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/pugixml.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/VTKFile.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/vtk_utils.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/XDMFFile.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_function.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/xdmf_mesh.cpp
