+++ cpp/dolfinx/io/VTKFile.cpp
@@ -7,6 +7,7 @@
 #include "VTKFile.h"
 #include "cells.h"
 #include "pugixml.hpp"
+#include "vtk_utils.h"
 #include "xdmf_utils.h"
 #include <dolfinx/common/IndexMap.h>
 #include <dolfinx/common/MPI.h>
@@ -34,187 +35,6 @@ namespace
 /// field
 constexpr std::array field_ext = {"_real", "_imag"};
 
-/// Tabulate the coordinate for every 'node' in a Lagrange function
-/// space.
-/// @param[in] V The function space. Must be a (discontinuous) Lagrange
-/// space.
-/// @return An array with shape (num_dofs, 3) array where the ith row
-/// corresponds to the coordinate of the ith dof in `V` (local to
-/// process)
-/// @pre `V` must be Lagrange and must not be a subspace
-xt::xtensor<double, 2>
-tabulate_lagrange_dof_coordinates(const dolfinx::fem::FunctionSpace& V)
-{
-  auto mesh = V.mesh();
-  assert(mesh);
-  const std::size_t gdim = mesh->geometry().dim();
-  const int tdim = mesh->topology().dim();
-
-  // Get dofmap data
-  auto dofmap = V.dofmap();
-  assert(dofmap);
-  auto map_dofs = dofmap->index_map;
-  assert(map_dofs);
-  const int index_map_bs = dofmap->index_map_bs();
-  const int dofmap_bs = dofmap->bs();
-
-  // Get element data
-  auto element = V.element();
-  assert(element);
-  const int e_bs = element->block_size();
-  const std::size_t scalar_dofs = element->space_dimension() / e_bs;
-  const std::int32_t num_nodes
-      = index_map_bs * (map_dofs->size_local() + map_dofs->num_ghosts())
-        / dofmap_bs;
-
-  // Get the dof coordinates on the reference element and the  mesh
-  // coordinate map
-  const xt::xtensor<double, 2>& X = element->interpolation_points();
-  const fem::CoordinateElement& cmap = mesh->geometry().cmap();
-
-  // Prepare cell geometry
-  const graph::AdjacencyList<std::int32_t>& dofmap_x
-      = mesh->geometry().dofmap();
-  xtl::span<const double> x_g = mesh->geometry().x();
-  const std::size_t num_dofs_g = dofmap_x.num_links(0);
-
-  xtl::span<const std::uint32_t> cell_info;
-  if (element->needs_dof_transformations())
-  {
-    mesh->topology_mutable().create_entity_permutations();
-    cell_info = xtl::span(mesh->topology().get_cell_permutation_info());
-  }
-  const auto apply_dof_transformation
-      = element->get_dof_transformation_function<double>();
-
-  // Tabulate basis functions at node reference coordinates
-  const xt::xtensor<double, 2> phi
-      = xt::view(cmap.tabulate(0, X), 0, xt::all(), xt::all(), 0);
-
-  // Loop over cells and tabulate dofs
-  auto map = mesh->topology().index_map(tdim);
-  assert(map);
-  const std::int32_t num_cells = map->size_local() + map->num_ghosts();
-  xt::xtensor<double, 2> x = xt::zeros<double>({scalar_dofs, gdim});
-  xt::xtensor<double, 2> coordinate_dofs({num_dofs_g, gdim});
-  xt::xtensor<double, 2> coords = xt::zeros<double>({num_nodes, 3});
-  for (std::int32_t c = 0; c < num_cells; ++c)
-  {
-    // Extract cell geometry
-    auto dofs_x = dofmap_x.links(c);
-    for (std::size_t i = 0; i < dofs_x.size(); ++i)
-    {
-      std::copy_n(std::next(x_g.begin(), 3 * dofs_x[i]), gdim,
-                  std::next(coordinate_dofs.begin(), i * gdim));
-    }
-
-    // Tabulate dof coordinates on cell
-    cmap.push_forward(x, coordinate_dofs, phi);
-    apply_dof_transformation(xtl::span(x.data(), x.size()),
-                             xtl::span(cell_info.data(), cell_info.size()), c,
-                             x.shape(1));
-
-    // Copy dof coordinates into vector
-    auto dofs = dofmap->cell_dofs(c);
-    for (std::size_t i = 0; i < dofs.size(); ++i)
-    {
-      std::int32_t dof = dofs[i];
-      for (std::size_t j = 0; j < gdim; ++j)
-        coords(dof, j) = x(i, j);
-    }
-  }
-
-  return coords;
-}
-//-----------------------------------------------------------------------------
-
-/// Given a FunctionSpace, create a topology and geometry based on the
-/// dof coordinates.
-/// @note Only supports (discontinuous) Lagrange functions
-/// @param[in] u The function
-std::pair<xt::xtensor<double, 2>, xt::xtensor<std::int64_t, 2>>
-vtk_mesh_from_space(const fem::FunctionSpace& V)
-{
-  auto mesh = V.mesh();
-  assert(mesh);
-  const int tdim = mesh->topology().dim();
-
-  assert(V.element());
-  if (V.element()->is_mixed())
-    throw std::runtime_error("Can create VTK mesh from a mixed element");
-
-  xt::xtensor<double, 2> x = tabulate_lagrange_dof_coordinates(V);
-  auto map = mesh->topology().index_map(tdim);
-  const std::size_t num_cells = map->size_local() + map->num_ghosts();
-
-  // Create permutation from DOLFINx dof ordering to VTK
-  auto dofmap = V.dofmap();
-  assert(dofmap);
-  const std::uint32_t num_nodes = dofmap->cell_dofs(0).size();
-  std::vector<std::uint8_t> vtkmap = dolfinx::io::cells::transpose(
-      io::cells::perm_vtk(mesh->topology().cell_type(), num_nodes));
-
-  // Extract topology for all local cells as
-  // [N0, v0_0, ...., v0_N0, N1, v1_0, ...., v1_N1, ....]
-  // std::vector<std::int64_t> vtk_topology(num_cells * (num_nodes + 1));
-  xt::xtensor<std::int64_t, 2> vtk_topology({num_cells, num_nodes});
-  for (std::size_t c = 0; c < num_cells; ++c)
-  {
-    auto dofs = dofmap->cell_dofs(c);
-    for (std::size_t i = 0; i < dofs.size(); ++i)
-      vtk_topology(c, i) = dofs[vtkmap[i]];
-  }
-
-  return {std::move(x), std::move(vtk_topology)};
-}
-//-----------------------------------------------------------------------------
-
-/// Extract the cell topology (connectivity) in VTK ordering for all
-/// cells the mesh. The 'topology' includes higher-order 'nodes'. The
-/// index of a 'node' corresponds to the index of DOLFINx geometry
-/// 'nodes'.
-/// @param [in] mesh The mesh
-/// @return The cell topology in VTK ordering and in term of the DOLFINx
-/// geometry 'nodes'
-/// @note The indices in the return array correspond to the point
-/// indices in the mesh geometry array
-/// @note Even if the indices are local (int32), both Fides and VTX
-/// require int64 as local input
-xt::xtensor<std::int64_t, 2> extract_vtk_connectivity(const mesh::Mesh& mesh)
-{
-  // Get DOLFINx to VTK permutation
-  // FIXME: Use better way to get number of nodes
-  const graph::AdjacencyList<std::int32_t>& dofmap_x = mesh.geometry().dofmap();
-  const std::size_t num_nodes = dofmap_x.num_links(0);
-  std::vector map = dolfinx::io::cells::transpose(
-      dolfinx::io::cells::perm_vtk(mesh.topology().cell_type(), num_nodes));
-  // TODO: Remove when when paraview issue 19433 is resolved
-  // (https://gitlab.kitware.com/paraview/paraview/issues/19433)
-  if (mesh.topology().cell_type() == dolfinx::mesh::CellType::hexahedron
-      and num_nodes == 27)
-  {
-    map = {0,  9, 12, 3,  1, 10, 13, 4,  18, 15, 21, 6,  19, 16,
-           22, 7, 2,  11, 5, 14, 8,  17, 20, 23, 24, 25, 26};
-  }
-  // Extract mesh 'nodes'
-  const int tdim = mesh.topology().dim();
-  const std::size_t num_cells = mesh.topology().index_map(tdim)->size_local();
-
-  // Build mesh connectivity
-
-  // Loop over cells
-  xt::xtensor<std::int64_t, 2> topology({num_cells, num_nodes});
-  for (std::size_t c = 0; c < num_cells; ++c)
-  {
-    // For each cell, get the 'nodes' and place in VTK order
-    auto dofs_x = dofmap_x.links(c);
-    for (std::size_t i = 0; i < dofs_x.size(); ++i)
-      topology(c, i) = dofs_x[map[i]];
-  }
-
-  return topology;
-}
-//-----------------------------------------------------------------------------
 //----------------------------------------------------------------------------
 /// Return true if Function is a cell-wise constant, otherwise false
 bool is_cellwise(const fem::FunctionSpace& V)
@@ -246,37 +66,6 @@ std::string get_counter(const pugi::xml_node& node, const std::string& name)
   const int num_digits = 6;
   std::string counter = std::to_string(n);
   return std::string(num_digits - counter.size(), '0').append(counter);
-}
-//----------------------------------------------------------------------------
-
-/// Get the VTK cell type integer
-std::int8_t get_vtk_cell_type(mesh::CellType cell, int dim)
-{
-  if (cell == mesh::CellType::prism and dim == 2)
-    throw std::runtime_error("More work needed for prism cell");
-
-  // Get cell type
-  mesh::CellType cell_type = mesh::cell_entity_type(cell, dim, 0);
-
-  // Determine VTK cell type (arbitrary Lagrange elements)
-  // https://vtk.org/doc/nightly/html/vtkCellType_8h_source.html
-  switch (cell_type)
-  {
-  case mesh::CellType::point:
-    return 1;
-  case mesh::CellType::interval:
-    return 68;
-  case mesh::CellType::triangle:
-    return 69;
-  case mesh::CellType::quadrilateral:
-    return 70;
-  case mesh::CellType::tetrahedron:
-    return 71;
-  case mesh::CellType::hexahedron:
-    return 72;
-  default:
-    throw std::runtime_error("Unknown cell type");
-  }
 }
 //----------------------------------------------------------------------------
 
@@ -389,7 +178,7 @@ void add_mesh(const xt::xtensor<double, 2>& x,
   type_node.append_attribute("type") = "Int8";
   type_node.append_attribute("Name") = "types";
   type_node.append_attribute("format") = "ascii";
-  int vtk_celltype = get_vtk_cell_type(celltype, tdim);
+  int vtk_celltype = io::get_vtk_cell_type(celltype, tdim);
   std::stringstream s;
   for (std::size_t c = 0; c < cells.shape(0); ++c)
     s << vtk_celltype << " ";
@@ -477,13 +266,13 @@ void write_function(
   xt::xtensor<std::int64_t, 2> cells;
   if (is_cellwise(*V0))
   {
-    cells = extract_vtk_connectivity(*mesh0);
+    cells = io::extract_vtk_connectivity(*mesh0);
     const mesh::Geometry& geometry = mesh0->geometry();
     x = xt::adapt(geometry.x().data(), geometry.x().size(), xt::no_ownership(),
                   std::vector({geometry.x().size() / 3, std::size_t(3)}));
   }
   else
-    std::tie(x, cells) = vtk_mesh_from_space(*V0);
+    std::tie(x, cells) = io::vtk_mesh_from_space(*V0);
 
   // Add "Piece" node and required metadata
   pugi::xml_node piece_node = grid_node_vtu.append_child("Piece");
