+++ python/test/unit/io/test_vtk.py
@@ -17,7 +17,6 @@ from dolfinx.mesh import (CellType, create_mesh, create_unit_cube,
                           create_unit_interval, create_unit_square)
 
 from mpi4py import MPI
-from petsc4py import PETSc
 
 cell_types_2D = [CellType.triangle, CellType.quadrilateral]
 cell_types_3D = [CellType.tetrahedron, CellType.hexahedron]
